Create README.md
Xcitement is around the corner!

Here at DriftXchange we love the Xfamily (XRP, XLM & XDC) like so many others. The difference here is we are intertwining all three of these powerhouses for investors who want to invest in all three at the same time.
​
From early next year you will be able to access the redeem feature on this site to redeem our fully backed DRIFTX token for an even balance of XRP, XLM & XDC.
​
We also aim to release a lot more features on this website like exchanging between XRP, XLM & XDC. Further on in 2022 we want to integrate all of their individual network tokens into one wallet for a maximum Xfamily Xperience.
 
This service has amazing value as it is backed by the greatest currencies through our own investments so you can rest assure that our DRIFT & DRIFTX tokens are backed a lot more than the average token.